added postman to its own snowpack app
