--- index.html
@@ -1,244 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="script.js" type="module"></script>
-    <title>Postman Clone</title>
-  </head>
-  <nav class="navbar navbar-dark bg-dark">
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-      <div class="container-fluid">
-        <a style="color: black" class="navbar-brand" href="./index.html"
-          >WebPulse</a
-        >
-        <button
-          class="navbar-toggler"
-          type="button"
-          data-bs-toggle="collapse"
-          data-bs-target="#navbarNavAltMarkup"
-          aria-controls="navbarNavAltMarkup"
-          aria-expanded="false"
-          aria-label="Toggle navigation"
-        >
-          <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-          <div class="navbar-nav">
-            <a
-              style="color: black"
-              class="nav-link active"
-              aria-current="page"
-              href="./index.html"
-              >Home</a
-            >
-            <a style="color: black" class="nav-link" href="#">About us</a>
-            <a style="color: black" class="nav-link" href="">IsItDown</a>
-            <a style="color: black" class="nav-link" href="./support"
-              >support</a
-            >
-          </div>
-        </div>
-      </div>
-    </nav>
-  </nav>
-  <body>
-    <div class="p-4">
-      <form data-form>
-        <div class="input-group mb-4">
-          <select class="form-select flex-grow-0 w-auto" data-method>
-            <option value="GET" selected>GET</option>
-            <option value="POST">POST</option>
-            <option value="PUT">PUT</option>
-            <option value="PATCH">PATCH</option>
-            <option value="DELETE">DELETE</option>
-          </select>
-          <input
-            data-url
-            required
-            class="form-control"
-            type="url"
-            placeholder="https://example.com"
-          />
-          <button type="submit" class="btn btn-primary">Send</button>
-        </div>
-
-        <ul class="nav nav-tabs" role="tablist">
-          <li class="nav-item" role="presentation">
-            <button
-              class="nav-link active"
-              id="query-params-tab"
-              data-bs-toggle="tab"
-              data-bs-target="#query-params"
-              type="button"
-              role="tab"
-              aria-controls="query-params"
-              aria-selected="true"
-            >
-              Query Params
-            </button>
-          </li>
-          <li class="nav-item" role="presentation">
-            <button
-              class="nav-link"
-              id="request-headers-tab"
-              data-bs-toggle="tab"
-              data-bs-target="#request-headers"
-              type="button"
-              role="tab"
-              aria-controls="request-headers"
-              aria-selected="false"
-            >
-              Headers
-            </button>
-          </li>
-          <li class="nav-item" role="presentation">
-            <button
-              class="nav-link"
-              id="json-tab"
-              data-bs-toggle="tab"
-              data-bs-target="#json"
-              type="button"
-              role="tab"
-              aria-controls="json"
-              aria-selected="false"
-            >
-              JSON
-            </button>
-          </li>
-        </ul>
-
-        <div class="tab-content p-3 border-top-0 border">
-          <div
-            class="tab-pane fade show active"
-            id="query-params"
-            role="tabpanel"
-            aria-labelledby="query-params-tab"
-          >
-            <div data-query-params></div>
-            <button
-              data-add-query-param-btn
-              class="mt-2 btn btn-outline-success"
-              type="button"
-            >
-              Add
-            </button>
-          </div>
-          <div
-            class="tab-pane fade"
-            id="request-headers"
-            role="tabpanel"
-            aria-labelledby="request-headers-tab"
-          >
-            <div data-request-headers></div>
-            <button
-              data-add-request-header-btn
-              class="mt-2 btn btn-outline-success"
-              type="button"
-            >
-              Add
-            </button>
-          </div>
-          <div
-            class="tab-pane fade"
-            id="json"
-            role="tabpanel"
-            aria-labelledby="json-tab"
-          >
-            <div
-              data-json-request-body
-              class="overflow-auto"
-              style="max-height: 200px"
-            ></div>
-          </div>
-        </div>
-      </form>
-
-      <div class="mt-5 d-none" data-response-section>
-        <h3>Response</h3>
-        <div class="d-flex my-2">
-          <div class="me-3">Status: <span data-status></span></div>
-          <div class="me-3">Time: <span data-time></span>ms</div>
-          <div class="me-3">Size: <span data-size></span></div>
-        </div>
-
-        <ul class="nav nav-tabs" role="tablist">
-          <li class="nav-item" role="presentation">
-            <button
-              class="nav-link active"
-              id="body-tab"
-              data-bs-toggle="tab"
-              data-bs-target="#body"
-              type="button"
-              role="tab"
-              aria-controls="body"
-              aria-selected="true"
-            >
-              Body
-            </button>
-          </li>
-          <li class="nav-item" role="presentation">
-            <button
-              class="nav-link"
-              id="response-headers-tab"
-              data-bs-toggle="tab"
-              data-bs-target="#response-headers"
-              type="button"
-              role="tab"
-              aria-controls="response-headers"
-              aria-selected="false"
-            >
-              Headers
-            </button>
-          </li>
-        </ul>
-
-        <div class="tab-content p-3 border-top-0 border">
-          <div
-            class="tab-pane fade show active"
-            id="body"
-            role="tabpanel"
-            aria-labelledby="body-tab"
-          >
-            <div
-              data-json-response-body
-              class="overflow-auto"
-              style="max-height: 200px"
-            ></div>
-          </div>
-          <div
-            class="tab-pane fade"
-            id="response-headers"
-            role="tabpanel"
-            aria-labelledby="response-headers-tab"
-          >
-            <div
-              style="
-                display: grid;
-                grid-template-columns: auto 1fr;
-                gap: 0.5rem 2rem;
-              "
-              data-response-headers
-            ></div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <template data-key-value-template>
-      <div class="input-group my-2" data-key-value-pair>
-        <input type="text" data-key class="form-control" placeholder="Key" />
-        <input
-          type="text"
-          data-value
-          class="form-control"
-          placeholder="Value"
-        />
-        <button type="button" data-remove-btn class="btn btn-outline-danger">
-          Remove
-        </button>
-      </div>
-    </template>
-  </body>
-</html>
