+++ views/webpulse.ejs
@@ -5,6 +5,42 @@
     <link
       rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
     />
+    <nav class="navbar navbar-dark bg-dark">
+        <nav class="navbar navbar-expand-lg navbar-light bg-light">
+          <div class="container-fluid">
+            <a style="color: black" class="navbar-brand" href="./index.html"
+              >WebPulse</a
+            >
+            <button
+              class="navbar-toggler"
+              type="button"
+              data-bs-toggle="collapse"
+              data-bs-target="#navbarNavAltMarkup"
+              aria-controls="navbarNavAltMarkup"
+              aria-expanded="false"
+              aria-label="Toggle navigation"
+            >
+              <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+              <div class="navbar-nav">
+                <a
+                  style="color: black"
+                  class="nav-link active"
+                  aria-current="page"
+                  href="./index.html"
+                  >Home</a
+                >
+                <a style="color: black" class="nav-link" href="#">About us</a>
+                <a style="color: black" class="nav-link" href="">IsItDown</a>
+                <a style="color: black" class="nav-link" href="./support"
+                  >support</a
+                >
+              </div>
+            </div>
+          </div>
+        </nav>
+      </nav>
   </head>
   <body>
       <div class="container">
